% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{coef.glmaag}
\alias{coef.glmaag}
\title{Coefficients for glmaag}
\usage{
\method{coef}{glmaag}(object, lam1, lam2, ...)
}
\arguments{
\item{object}{fitted glmaag object}

\item{lam1}{lambda1 sequence need coefficients, must be within the fitted model}

\item{lam2}{lambda2 sequence need coefficients, must be within the fitted model}

\item{...}{\dots}
}
\value{
coefficients
}
\description{
Get coefficients for glmaag objects
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, -(1:3)]
mod <- glmaag(y, x, L0)
cc <- coef(mod)
}
