% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{coef.ss_glmaag}
\alias{coef.ss_glmaag}
\title{Coefficients for ss_glmaag}
\usage{
\method{coef}{ss_glmaag}(object, ...)
}
\arguments{
\item{object}{the model estimated via stability selection}

\item{...}{\dots}
}
\value{
the optimal coefficients get from stability selection including intercept (except for Cox)
}
\description{
Get the coefficients tuned by stability selection
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, -(1:3)]
mod <- ss_glmaag(y, x, L0, nsam = 3)
cc <- coef(mod)
}
