% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcn.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate prediction}
\usage{
evaluate(y_pre, y, cutpoint = 0.5, fam = "Gaussian")
}
\arguments{
\item{y_pre}{predicted value}

\item{y}{actual values (class for binary phenotype and Surv object for right censored phenotype)}

\item{cutpoint}{cutpoints for binary phenotype, default to be 0.5}

\item{fam}{family of the phenotype, can be "continous", "binary", or "Cox"}
}
\value{
goodness of prediction
}
\description{
Evaluate goodness of prediction.
}
\examples{
x <- rnorm(100)
y <- rnorm(100)
evaluate(x, y)
}
