% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{L1}
\alias{L1}
\title{sample network 1}
\format{a matrix with 20 rows and 20 columns}
\usage{
L1
}
\description{
An alternative Laplacian matrix for the predictors
}
\keyword{datasets}
