% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcn.R
\name{plot.cv_glmaag}
\alias{plot.cv_glmaag}
\title{Cross validation plot}
\usage{
\method{plot}{cv_glmaag}(x, col_count = 3, SE = T, ...)
}
\arguments{
\item{x}{the cv_glmaag object}

\item{col_count}{number of columns in the plots}

\item{SE}{whether or not plot the standared error curves (when SE = TRUE)}

\item{...}{\dots}
}
\value{
plot generated by the model
}
\description{
plot cross validation performance paths
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, -(1:3)]
cvwhich <- sample(rep(0:4, length.out = length(y)))
mod <- cv_glmaag(y, x, L0, cvwhich = cvwhich)
gg <- plot(mod, SE = FALSE)
}
