% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{plot.glmaag}
\alias{plot.glmaag}
\title{Paths for glmaag object}
\usage{
\method{plot}{glmaag}(x, col_count = 3, type = "coef", ...)
}
\arguments{
\item{x}{glmaag object}

\item{col_count}{number of columns shown in the plot (when type = 'coef)}

\item{type}{can be "coef" (coefficients paths), "loglik" (log likelihood paths), or "n" (number of parameters paths)}

\item{...}{\dots}
}
\value{
plots
}
\description{
Generates coefficients, lok likelihood, or number of parameters paths for glmaag models
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, -(1:3)]
mod <- glmaag(y, x, L0)
gg <- plot(mod, type = 'loglik')
}
