% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcn.R
\name{predict.cv_glmaag}
\alias{predict.cv_glmaag}
\title{Predict}
\usage{
\method{predict}{cv_glmaag}(object, x, type1se = T, type = "link", ...)
}
\arguments{
\item{object}{the estimated cv_glmaag model}

\item{x}{the new dataset for prediction, if omitted returns the training prediction}

\item{type1se}{whether or not using the coefficients by one standard error ruld, default to be TRUE}

\item{type}{can be either "link", or "response", link returns linear predicted score, For Gaussian model this option can be ingnored, for logistic model "response" returns predicted probability, for Cox model "reponse" returns relative risk}

\item{...}{\dots}
}
\value{
the predicted value
}
\description{
Prediction for cv_glmaag model
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, -(1:3)]
cvwhich <- sample(rep(0:4, length.out = length(y)))
mod <- cv_glmaag(y, x, L0, cvwhich = cvwhich)
pp <- predict(mod)
}
