% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcn.R
\name{print.cv_glmaag}
\alias{print.cv_glmaag}
\title{the results of the cross validation model}
\usage{
\method{print}{cv_glmaag}(x, ...)
}
\arguments{
\item{x}{the fitted cv_glmaag object}

\item{...}{\dots}
}
\description{
print fitted information
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, -(1:3)]
cvwhich <- sample(rep(0:4, length.out = length(y)))
mod <- cv_glmaag(y, x, L0, cvwhich = cvwhich)
print(mod)
}
