% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{print.ss_glmaag}
\alias{print.ss_glmaag}
\title{the results of the stability selection model}
\usage{
\method{print}{ss_glmaag}(x, ...)
}
\arguments{
\item{x}{the fitted ss_glmaag object}

\item{...}{\dots}
}
\description{
print fitted information
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, -(1:3)]
mod <- ss_glmaag(y, x, L0, nsam = 3)
print(mod)
}
