% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runtheExample}
\alias{runtheExample}
\title{Shiny app}
\usage{
runtheExample(whetherrun)
}
\arguments{
\item{whetherrun}{whether to run shiny app, default to be TRUE}
}
\description{
Run the shiny app for web interative using. You need to load data.table, dplyr, ggplot2, plotROC, and survminer beforehand. If you want to do parallel computing, you also need to register cores.
}
\examples{
runtheExample(FALSE)
}
