% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampledata}
\alias{sampledata}
\title{Simulated data}
\format{a data frame with 100 rows and 23 variables}
\usage{
sampledata
}
\description{
A data set containing outcome for Gaussian, logistic, and Cox variables and 20 predictors
}
\keyword{datasets}
