% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{plot.ss_glmaag}
\alias{plot.ss_glmaag}
\title{Instability plot}
\usage{
\method{plot}{ss_glmaag}(x, ...)
}
\arguments{
\item{x}{the input ss_glmagrph object}

\item{...}{\dots}
}
\value{
the instability path
}
\description{
Instability path plot
}
\examples{
data(sampledata)
data(L0)
y <- sampledata$Y_Gau
x <- sampledata[, 4:6]
mod <- ss_glmaag(y, x, L0[seq_len(3), seq_len(3)], nsam = 3)
gg <- plot(mod)
}
